requirements.txt
uzu-accounts-app==2.0.0
python-dotenv==0.15.0
Django==2.0.6
mysqlclient==2.0.2